+++ CHANGELOG.md
@@ -167,7 +167,7 @@
 - Added `craft\events\TransformImageEvent`.
 - Added `craft\fieldlayoutelements\AddressField`.
 - Added `craft\fieldlayoutelements\AddressLatitudeLongitudeField`.
-- Added `craft\fieldlayoutelements\AssetAltField`.
+- Added `craft\fieldlayoutelements\assets\AltField`.
 - Added `craft\fieldlayoutelements\BaseNativeField`, which replaces `craft\fieldlayoutelements\StandardField`.
 - Added `craft\fieldlayoutelements\TextareaField`.
 - Added `craft\fieldlayoutelements\TextField`, which replaces `craft\fieldlayoutelements\StandardTextField`.
@@ -530,6 +530,8 @@
 - `craft\events\GetAssetThumbUrlEvent` has been renamed to `DefineAssetThumbUrlEvent`.
 - `craft\events\GetAssetUrlEvent` has been renamed to `DefineAssetUrlEvent`.
 - `craft\events\RevisionEvent::$source` has been renamed to `$canonical`.
+- `craft\fieldlayoutelements\AssetTitleField` has been renomed to `craft\fieldlayoutelements\assets\AssetTitleField`.
+- `craft\fieldlayoutelements\EntryTitleField` has been renomed to `craft\fieldlayoutelements\entries\EntryTitleField`.
 - `craft\fields\BaseRelationField::$limit` has been renamed to `$maxRelations`.
 - `craft\fields\BaseRelationField::elementType()` is now public.
 - `craft\fields\BaseRelationField::inputSelectionCriteria()` has been renamed to `getInputSelectionCriteria()`, and is now public.